The WAR of the WaRnINGS, Part the 5th: All tests passed.
git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@52677 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: a430222f4b083282a0df881d678607f6a3989bac

+++ rosetta_source/src/protocols/forge/methods/util.cc
@@ -347,7 +347,7 @@ parse_resfile_string_with_no_lockdown( core::pose::Pose const & pose, core::pack
 			if(pose.pdb_info()){
 				resid = pose.pdb_info()->pdb2pose().find( chain, PDBnum, icode );
 			} else {
-				if((1 <= PDBnum) && (PDBnum <= pose.total_residue())) {
+				if((1 <= PDBnum) && (PDBnum <= (int)pose.total_residue())) {
 					resid = PDBnum;
 				}
 			}
